Tweak generators (#69)
