--- bin/ruby_test
@@ -1,17 +1,0 @@
-#!/usr/bin/env ruby
-$: << File.expand_path(File.expand_path("../../test", __FILE__))
-
-# Remove this begin/rescue once Rails 4 support is removed.
-begin
-  require "bundler/setup"
-  require "rails/test_unit/runner"
-  require "rails/test_unit/reporter"
-  require "rails/test_unit/line_filtering"
-
-  Rails::TestUnitReporter.executable = "bin/test"
-
-  Rails::TestUnit::Runner.parse_options(ARGV)
-  Rails::TestUnit::Runner.run(ARGV)
-rescue LoadError
-  exec "rake"
-end
