--- lib/generators/stimulus_reflex/USAGE
@@ -1,23 +1,0 @@
-Description:
-    Generate boilerplate files to get you up and running with stimulus_reflex
-
-Example:
-    rails generate stimulus_reflex example
-
-    This will create
-        app/javascript/controllers/application_controller.js
-        app/javascript/controllers/example_controller.js
-        app/reflexes/application_reflex.rb
-        app/reflexes/example_reflex.rb
-
-    rails generate stimulus_reflex:controller example
-
-    This will create:
-        app/javascript/controllers/application_controller.js (if it does not already exist)
-        app/javascript/controllers/example_controller.js
-
-    rails generate stimulus_reflex:reflex example
-
-    This will create:
-        app/reflexes/application_reflex.rb (if it does not already exist)
-        app/reflexes/example_reflex.rb
