--- lib/generators/stimulus_reflex/templates/application_controller.js
@@ -1,24 +1,0 @@
-import { Controller } from 'stimulus'
-import StimulusReflex from 'stimulus_reflex'
-
-export default class extends Controller {
-  connect () {
-    StimulusReflex.register(this)
-  }
-
-  beforeReflex (element, reflex) {
-    // do stuff
-  }
-
-  afterReflex (element, reflex) {
-    // do stuff
-  }
-
-  reflexError (element, reflex, error) {
-    // do stuff
-  }
-
-  reflexSuccess (element, reflex, error) {
-    // do stuff
-  }
-}
