--- lib/generators/stimulus_reflex/templates/application_reflex.rb
@@ -1,2 +1,0 @@
-class ApplicationReflex < StimulusReflex::Reflex
-end
