--- lib/generators/stimulus_reflex/templates/example_controller.js
@@ -1,19 +1,0 @@
-import ApplicationController from './application_controller'
-
-export default class extends ApplicationController {
-  // beforeUpdate(anchorElement) {
-  //  anchorElement.innerText = 'Updating...'
-  // }
-
-  // beforeDelete(anchorElement) {
-  //   anchorElement.innerText = 'Deleting...'
-  // }
-
-  // afterUpdate(anchorElement) {
-  //   anchorElement.innerText = 'Updated!'
-  // }
-
-  // afterDelete(anchorElement) {
-  //   anchorElement.innerText = 'Updated!'
-  // }
-}
