--- lib/generators/stimulus_reflex/templates/example_reflex.rb.erb
@@ -1,2 +1,0 @@
-class <%= @name.capitalize %>Reflex < ApplicationReflex
-end
