+++ lib/generators/templates/application_reflex.rb
@@ -1,0 +1,10 @@
+class ApplicationReflex < StimulusReflex::Reflex
+  # Put application wide Reflex behavior in this file.
+  #
+  # Example:
+  #
+  #   # If your ActionCable connection is: `identified_by :current_user`
+  #   delegate :current_user, to: :channel
+  #
+  # Learn more at: https://docs.stimulusreflex.com
+end
