+++ lib/generators/templates/custom_reflex.rb
@@ -1,0 +1,21 @@
+class <%= @name.classify %>Reflex < ApplicationReflex
+  # Add Reflex methods in this file.
+  #
+  # All Reflex instances expose the following properties:
+  #
+  #   - connection - the ActionCable connection
+  #   - channel - the ActionCable channel
+  #   - request - an ActionDispatch::Request proxy for the socket connection
+  #   - session - the ActionDispatch::Session store for the current visitor
+  #   - url - the URL of the page that triggered the reflex
+  #   - element - a Hash like object that represents the HTML element that triggered the reflex
+  #
+  # Example:
+  #
+  #   def example(argument=true)
+  #     # Your logic here...
+  #     # Any declared instance variables will be made available to the Rails controller and view.
+  #   end
+  #
+  # Learn more at: https://docs.stimulusreflex.com
+end
