+++ lib/stimulus_reflex/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module StimulusReflex
-  VERSION = "2.1.2"
+  VERSION = "2.1.3"
 end
