+++ test/test_helper.rb
@@ -1,5 +1,2 @@
-# For generators
-require "rails/generators/test_case"
-require "generators/stimulus_reflex/stimulus_reflex_generator"
-require "generators/stimulus_reflex/reflex_generator"
-require "generators/stimulus_reflex/controller_generator"
+ENV["RAILS_ENV"] ||= "test"
+require_relative "../lib/stimulus_reflex"
